extract test to dedicated class
