+++ matrix-sdk-android/src/androidTest/java/org/matrix/android/sdk/account/AccountCreationTest.kt
@@ -16,17 +16,7 @@
 
 package org.matrix.android.sdk.account
 
-import android.util.Log
 import androidx.test.filters.LargeTest
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Deferred
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.asCoroutineDispatcher
-import kotlinx.coroutines.async
-import kotlinx.coroutines.awaitAll
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 import org.junit.FixMethodOrder
 import org.junit.Ignore
 import org.junit.Test
@@ -38,9 +28,6 @@ import org.matrix.android.sdk.common.CommonTestHelper
 import org.matrix.android.sdk.common.CryptoTestHelper
 import org.matrix.android.sdk.common.SessionTestParams
 import org.matrix.android.sdk.common.TestConstants
-import java.util.concurrent.Executors
-import java.util.concurrent.atomic.AtomicBoolean
-import kotlin.random.Random
 
 @RunWith(JUnit4::class)
 @FixMethodOrder(MethodSorters.JVM)
@@ -74,104 +61,5 @@ class AccountCreationTest : InstrumentedTest {
         val res = cryptoTestHelper.doE2ETestWithAliceInARoom()
 
         res.cleanUp(commonTestHelper)
-    }
-
-    @Test
-    fun testConcurrentDecrypt() {
-//        val res = cryptoTestHelper.doE2ETestWithAliceAndBobInARoom()
-
-        // =============================
-        // ARRANGE
-        // =============================
-
-        val aliceSession = commonTestHelper.createAccount(TestConstants.USER_ALICE, SessionTestParams(true))
-        val bobSession = commonTestHelper.createAccount(TestConstants.USER_BOB, SessionTestParams(true))
-        cryptoTestHelper.initializeCrossSigning(bobSession)
-        val bobSession2 = commonTestHelper.logIntoAccount(bobSession.myUserId, SessionTestParams(true))
-
-        bobSession2.cryptoService().verificationService().markedLocallyAsManuallyVerified(bobSession.myUserId, bobSession.sessionParams.deviceId ?: "")
-        bobSession.cryptoService().verificationService().markedLocallyAsManuallyVerified(bobSession.myUserId, bobSession2.sessionParams.deviceId ?: "")
-
-        val roomId = cryptoTestHelper.createDM(aliceSession, bobSession)
-        val roomAlicePOV = aliceSession.getRoom(roomId)!!
-
-        // =============================
-        // ACT
-        // =============================
-
-        val timelineEvent = commonTestHelper.sendTextMessage(roomAlicePOV, "Hello Bob", 1).first()
-        val secondEvent = commonTestHelper.sendTextMessage(roomAlicePOV, "Hello Bob 2", 1).first()
-        val thirdEvent = commonTestHelper.sendTextMessage(roomAlicePOV, "Hello Bob 3", 1).first()
-        val forthEvent = commonTestHelper.sendTextMessage(roomAlicePOV, "Hello Bob 4", 1).first()
-
-        // await for bob unverified session to get the message
-        commonTestHelper.waitWithLatch { latch ->
-            commonTestHelper.retryPeriodicallyWithLatch(latch) {
-                bobSession.getRoom(roomId)?.getTimeLineEvent(forthEvent.eventId) != null
-            }
-        }
-
-        val eventBobPOV = bobSession.getRoom(roomId)?.getTimeLineEvent(timelineEvent.eventId)!!
-        val secondEventBobPOV = bobSession.getRoom(roomId)?.getTimeLineEvent(secondEvent.eventId)!!
-        val thirdEventBobPOV = bobSession.getRoom(roomId)?.getTimeLineEvent(thirdEvent.eventId)!!
-        val forthEventBobPOV = bobSession.getRoom(roomId)?.getTimeLineEvent(forthEvent.eventId)!!
-
-        // let's try to decrypt concurrently and check that we are not getting exceptions
-        val dispatcher = Executors
-                .newFixedThreadPool(100)
-                .asCoroutineDispatcher()
-        val coroutineScope = CoroutineScope(SupervisorJob() + dispatcher)
-
-        val eventList = listOf(eventBobPOV, secondEventBobPOV, thirdEventBobPOV, forthEventBobPOV)
-
-        val atomicAsError = AtomicBoolean()
-        val deff = mutableListOf<Deferred<Any>>()
-
-        val cryptoService = bobSession.cryptoService()
-
-        coroutineScope.launch {
-            for (spawn in 1..100) {
-                delay((Random.nextFloat() * 1000).toLong())
-                aliceSession.cryptoService().requestRoomKeyForEvent(eventList.random().root)
-            }
-        }
-
-        for (spawn in 1..8000) {
-            eventList.random().let { event ->
-                coroutineScope.async {
-                    try {
-                        cryptoService.decryptEvent(event.root, "")
-                        Log.d("#TEST", "[$spawn] Decrypt Success ${event.eventId} :${Thread.currentThread().name}")
-                    } catch (failure: Throwable) {
-                        atomicAsError.set(true)
-                        Log.e("#TEST", "Failed to decrypt $spawn/${event.eventId}  :$failure")
-                    }
-                }.let {
-                    deff.add(it)
-                }
-            }
-        }
-
-        coroutineScope.launch {
-            for (spawn in 1..100) {
-                delay((Random.nextFloat() * 1000).toLong())
-                bobSession.cryptoService().requestRoomKeyForEvent(eventList.random().root)
-            }
-        }
-
-        commonTestHelper.runBlockingTest(10 * 60_000) {
-            deff.awaitAll()
-            delay(10_000)
-            assert(!atomicAsError.get())
-            // There should be no errors
-        }
-
-
-
-        coroutineScope.cancel()
-
-        commonTestHelper.signOutAndClose(aliceSession)
-        commonTestHelper.signOutAndClose(bobSession)
-        commonTestHelper.signOutAndClose(bobSession2)
     }
 }
